Instructions for new github host
Inscructions in README.md reflects git clone of the original project which is no longer on the internet. Updated README file with instructions for the new link.
Directory of rubycoind is RubyCoin/src whereas rubycoin/src is a different directory. This change is important for users who are basically just copying and pasting commands into their terminal